[BAEL-14088] - Moved code for RequestBody and ResponseBody annotations article
