+++ spring-boot-rest/src/main/java/com/baeldung/interceptors/RestTemplateHeaderModifierInterceptor.java
@@ -1,4 +1,4 @@
-package com.baeldung.sampleapp.interceptors;
+package com.baeldung.interceptors;
 
 import java.io.IOException;
