+++ spring-boot-rest/src/main/java/com/baeldung/services/ExampleService.java
