+++ spring-boot-rest/src/main/java/com/baeldung/transfer/ResponseTransfer.java
