+++ spring-boot-rest/src/main/java/com/baeldung/web/controller/ExamplePostController.java
@@ -1,4 +1,4 @@
-package com.baeldung.controllers;
+package com.baeldung.web.controller;
 
 import com.baeldung.services.ExampleService;
 import com.baeldung.transfer.ResponseTransfer;
